Fit Bivariate Copula Model with Optional Marginal Parameter Estimation

The `fit_bivariate_copula` function fits a bivariate copula model to data with optional marginal parameter estimation. It supports Gaussian, Clayton, and Joe copulas, and allows for specifying normal, lognormal, exponential, or beta marginal distributions for each variable. This function utilizes Stan for Bayesian inference.

Usage

fit_bivariate_copula(
  U,
  copula,
  marginals,
  iter = 2000,
  chains = 4,
  warmup = 1000,
  thin = 1,
  seed = NULL,
  control = list(adapt_delta = 0.8, max_treedepth = 10),
  cores = 1
)

Arguments

U

Data matrix of observed marginals, with two variables.

copula

The type of copula to fit. Options are "gaussian", "clayton", and "joe".

marginals

A list specifying the marginal distributions for each variable. Options are "normal", "lognormal", "exponential", and "beta".

iter

Number of iterations for each chain. Default is 2000.

chains

Number of chains. Default is 4.

warmup

Number of warmup iterations per chain. Default is 1000.

thin

Thinning rate. Default is 1.

seed

Random seed. Default is NULL.

control

A list of parameters to control the sampler's behavior. Default is list(adapt_delta = 0.8, max_treedepth = 10).

cores

Number of cores to use for parallel processing. Default is 1.

Value

A list containing:

fit

Stan model object.

Examples

if (FALSE) { # \dontrun{
library(copulaStan)
library(copula)

seed <- 2024
set.seed(seed)
true_rho <- 0.5
n <- 2000

margins <- c("norm", "lnorm")
params <- list(list(mean = 0.8, sd = 2), list(meanlog = 0, sdlog = 0.8))
cop <- normalCopula(param = true_rho, dim = 2)
mvdc_copula <- mvdc(cop, margins = margins, paramMargins = params)
data <- rMvdc(n, mvdc_copula)

fit <- fit_bivariate_copula(data,
                            copula = "gaussian", marginals = c("normal", "lognormal"),
                            seed = seed)
} # }